factor out queue stuff part 1
