+++ common/persistence/sql/sqlplugin/matching_task_queue.go
@@ -3,9 +3,13 @@ package sqlplugin
 import (
 	"context"
 	"database/sql"
+	"strings"
+	"sync"
 )
 
 type (
+	MatchingTaskVersion int
+
 	// TaskQueuesRow represents a row in task_queues table
 	TaskQueuesRow struct {
 		RangeHash    uint32
@@ -35,16 +39,38 @@ type (
 		DataEncoding  string
 	}
 
-	// MatchingTaskQueue is the SQL persistence interface for matching task queues
+	// MatchingTaskQueue is the SQL persistence interface for matching task queues.
+	// This handles both "v1" and "v2" tables so that we don't have to duplicate as much code.
 	MatchingTaskQueue interface {
-		InsertIntoTaskQueues(ctx context.Context, row *TaskQueuesRow) (sql.Result, error)
-		UpdateTaskQueues(ctx context.Context, row *TaskQueuesRow) (sql.Result, error)
+		InsertIntoTaskQueues(ctx context.Context, row *TaskQueuesRow, v MatchingTaskVersion) (sql.Result, error)
+		UpdateTaskQueues(ctx context.Context, row *TaskQueuesRow, v MatchingTaskVersion) (sql.Result, error)
 		// SelectFromTaskQueues returns one or more rows from task_queues table
 		// Required Filter params:
 		//  to read a single row: {shardID, namespaceID, name, taskType}
 		//  to range read multiple rows: {shardID, namespaceIDGreaterThan, nameGreaterThan, taskTypeGreaterThan, pageSize}
-		SelectFromTaskQueues(ctx context.Context, filter TaskQueuesFilter) ([]TaskQueuesRow, error)
-		DeleteFromTaskQueues(ctx context.Context, filter TaskQueuesFilter) (sql.Result, error)
-		LockTaskQueues(ctx context.Context, filter TaskQueuesFilter) (int64, error)
+		SelectFromTaskQueues(ctx context.Context, filter TaskQueuesFilter, v MatchingTaskVersion) ([]TaskQueuesRow, error)
+		DeleteFromTaskQueues(ctx context.Context, filter TaskQueuesFilter, v MatchingTaskVersion) (sql.Result, error)
+		LockTaskQueues(ctx context.Context, filter TaskQueuesFilter, v MatchingTaskVersion) (int64, error)
 	}
 )
+
+const (
+	MatchingTaskVersion1 MatchingTaskVersion = 1
+	MatchingTaskVersion2 MatchingTaskVersion = 2
+)
+
+var switchTaskQueuesTableV2Cache sync.Map
+
+func SwitchTaskQueuesTable(baseQuery string, v MatchingTaskVersion) string {
+	if v == MatchingTaskVersion1 {
+		return baseQuery
+	} else if v != MatchingTaskVersion2 {
+		return "_invalid_version_"
+	}
+	if v2query, ok := switchTaskQueuesTableV2Cache.Load(baseQuery); ok {
+		return v2query.(string)
+	}
+	v2query := strings.ReplaceAll(baseQuery, " task_queues ", " task_queues_v2 ")
+	switchTaskQueuesTableV2Cache.Store(baseQuery, v2query)
+	return v2query
+}
