+++ common/persistence/sql/task_v2.go
@@ -1,7 +1,6 @@
 package sql
 
 import (
-	"bytes"
 	"context"
 	"database/sql"
 	"fmt"
@@ -18,256 +17,24 @@ import (
 type (
 	sqlTaskManagerV2 struct {
 		SqlStore
-		taskScanPartitions uint32
 		userDataStore
+		taskQueueStore
 	}
 )
 
-func newTaskManagerV2(db sqlplugin.DB, uds *userDataStore, taskScanPartitions int, logger log.Logger) (*sqlTaskManagerV2, error) {
+func newTaskManagerV2(
+	db sqlplugin.DB,
+	uds userDataStore,
+	tqs taskQueueStore,
+	logger log.Logger,
+) (*sqlTaskManagerV2, error) {
 	return &sqlTaskManagerV2{
-		SqlStore:           NewSqlStore(db, logger),
-		taskScanPartitions: uint32(taskScanPartitions),
-		userDataStore:      *uds,
+		SqlStore:       NewSqlStore(db, logger),
+		userDataStore:  uds,
+		taskQueueStore: tqs,
 	}, nil
 }
 
-func (m *sqlTaskManagerV2) CreateTaskQueue(
-	ctx context.Context,
-	request *persistence.InternalCreateTaskQueueRequest,
-) error {
-	nidBytes, err := primitives.ParseUUID(request.NamespaceID)
-	if err != nil {
-		return serviceerror.NewInternal(err.Error())
-	}
-	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
-
-	row := sqlplugin.TaskQueuesRowV2{
-		RangeHash:    tqHash,
-		TaskQueueID:  tqId,
-		RangeID:      request.RangeID,
-		Data:         request.TaskQueueInfo.Data,
-		DataEncoding: request.TaskQueueInfo.EncodingType.String(),
-	}
-	if _, err := m.Db.InsertIntoTaskQueuesV2(ctx, &row); err != nil {
-		if m.Db.IsDupEntryError(err) {
-			return &persistence.ConditionFailedError{Msg: err.Error()}
-		}
-		return serviceerror.NewUnavailablef("CreateTaskQueue operation failed. Failed to make task queue %v of type %v. Error: %v", request.TaskQueue, request.TaskType, err)
-	}
-
-	return nil
-}
-
-func (m *sqlTaskManagerV2) GetTaskQueue(
-	ctx context.Context,
-	request *persistence.InternalGetTaskQueueRequest,
-) (*persistence.InternalGetTaskQueueResponse, error) {
-	nidBytes, err := primitives.ParseUUID(request.NamespaceID)
-	if err != nil {
-		return nil, serviceerror.NewInternal(err.Error())
-	}
-	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
-	rows, err := m.Db.SelectFromTaskQueuesV2(ctx, sqlplugin.TaskQueuesFilterV2{
-		RangeHash:   tqHash,
-		TaskQueueID: tqId,
-	})
-
-	switch err {
-	case nil:
-		if len(rows) != 1 {
-			return nil, serviceerror.NewUnavailablef(
-				"GetTaskQueue operation failed. Expect exactly one result row, but got %d for task queue %v of type %v",
-				len(rows), request.TaskQueue, request.TaskType)
-		}
-		row := rows[0]
-		return &persistence.InternalGetTaskQueueResponse{
-			RangeID:       row.RangeID,
-			TaskQueueInfo: persistence.NewDataBlob(row.Data, row.DataEncoding),
-		}, nil
-	case sql.ErrNoRows:
-		return nil, serviceerror.NewNotFoundf(
-			"GetTaskQueue operation failed. TaskQueue: %v, TaskQueueType: %v, Error: %v",
-			request.TaskQueue, request.TaskType, err)
-	default:
-		return nil, serviceerror.NewUnavailablef(
-			"GetTaskQueue operation failed. Failed to check if task queue %v of type %v existed. Error: %v",
-			request.TaskQueue, request.TaskType, err)
-	}
-}
-
-func (m *sqlTaskManagerV2) UpdateTaskQueue(
-	ctx context.Context,
-	request *persistence.InternalUpdateTaskQueueRequest,
-) (*persistence.UpdateTaskQueueResponse, error) {
-	nidBytes, err := primitives.ParseUUID(request.NamespaceID)
-	if err != nil {
-		return nil, serviceerror.NewInternal(err.Error())
-	}
-
-	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue, request.TaskType, persistence.SubqueueZero)
-	var resp *persistence.UpdateTaskQueueResponse
-	err = m.SqlStore.txExecute(ctx, "UpdateTaskQueue", func(tx sqlplugin.Tx) error {
-		if err := lockTaskQueueV2(ctx,
-			tx,
-			tqHash,
-			tqId,
-			request.PrevRangeID,
-		); err != nil {
-			return err
-		}
-		result, err := tx.UpdateTaskQueuesV2(ctx, &sqlplugin.TaskQueuesRowV2{
-			RangeHash:    tqHash,
-			TaskQueueID:  tqId,
-			RangeID:      request.RangeID,
-			Data:         request.TaskQueueInfo.Data,
-			DataEncoding: request.TaskQueueInfo.EncodingType.String(),
-		})
-		if err != nil {
-			return err
-		}
-		rowsAffected, err := result.RowsAffected()
-		if err != nil {
-			return err
-		}
-		if rowsAffected != 1 {
-			return fmt.Errorf("%v rows were affected instead of 1", rowsAffected)
-		}
-		resp = &persistence.UpdateTaskQueueResponse{}
-		return nil
-	})
-	return resp, err
-}
-
-func (m *sqlTaskManagerV2) ListTaskQueue(
-	ctx context.Context,
-	request *persistence.ListTaskQueueRequest,
-) (*persistence.InternalListTaskQueueResponse, error) {
-	pageToken := taskQueuePageToken{MinTaskQueueId: minTaskQueueId}
-	if request.PageToken != nil {
-		if err := gobDeserialize(request.PageToken, &pageToken); err != nil {
-			return nil, serviceerror.NewInternalf("error deserializing page token: %v", err)
-		}
-	}
-	var err error
-	var rows []sqlplugin.TaskQueuesRowV2
-	var shardGreaterThan uint32
-	var shardLessThan uint32
-
-	i := uint32(0)
-	if pageToken.MinRangeHash > 0 {
-		// Resume partition position from page token, if exists, before entering loop
-		i = getPartitionForRangeHash(pageToken.MinRangeHash, m.taskScanPartitions)
-	}
-
-	lastPageFull := !bytes.Equal(pageToken.MinTaskQueueId, minTaskQueueId)
-	for ; i < m.taskScanPartitions; i++ {
-		// Get start/end boundaries for partition
-		shardGreaterThan, shardLessThan = getBoundariesForPartition(i, m.taskScanPartitions)
-
-		// If page token hash is greater than the boundaries for this partition, use the pageToken hash for resume point
-		if pageToken.MinRangeHash > shardGreaterThan {
-			shardGreaterThan = pageToken.MinRangeHash
-		}
-
-		filter := sqlplugin.TaskQueuesFilterV2{
-			RangeHashGreaterThanEqualTo: shardGreaterThan,
-			RangeHashLessThanEqualTo:    shardLessThan,
-			TaskQueueIDGreaterThan:      minTaskQueueId,
-			PageSize:                    &request.PageSize,
-		}
-
-		if lastPageFull {
-			// Use page token TaskQueueID filter for this query and set this to false
-			// in order for the next partition so we don't miss any results.
-			filter.TaskQueueIDGreaterThan = pageToken.MinTaskQueueId
-			lastPageFull = false
-		}
-
-		rows, err = m.Db.SelectFromTaskQueuesV2(ctx, filter)
-		if err != nil {
-			return nil, serviceerror.NewUnavailable(err.Error())
-		}
-
-		if len(rows) > 0 {
-			break
-		}
-	}
-
-	maxRangeHash := uint32(0)
-	resp := &persistence.InternalListTaskQueueResponse{
-		Items: make([]*persistence.InternalListTaskQueueItem, len(rows)),
-	}
-
-	for i, row := range rows {
-		resp.Items[i] = &persistence.InternalListTaskQueueItem{
-			RangeID:   row.RangeID,
-			TaskQueue: persistence.NewDataBlob(row.Data, row.DataEncoding),
-		}
-
-		// Only want to look at up to PageSize number of records to prevent losing data.
-		if row.RangeHash > maxRangeHash {
-			maxRangeHash = row.RangeHash
-		}
-
-		// Enforces PageSize
-		if i >= request.PageSize-1 {
-			break
-		}
-	}
-
-	var nextPageToken []byte
-	switch {
-	case len(rows) >= request.PageSize:
-		// Store the details of the lastRow seen up to PageSize.
-		// Note we don't increment the rangeHash as we do in the case below.
-		// This is so we can exhaust this hash before moving forward.
-		lastRow := &rows[request.PageSize-1]
-		nextPageToken, err = gobSerialize(&taskQueuePageToken{
-			MinRangeHash:   shardGreaterThan,
-			MinTaskQueueId: lastRow.TaskQueueID,
-		})
-	case shardLessThan < math.MaxUint32:
-		// Create page token with +1 from the last rangeHash we have seen to prevent duplicating the last row.
-		// Since we have not exceeded PageSize, we are confident we won't lose data here and we have exhausted this hash.
-		nextPageToken, err = gobSerialize(&taskQueuePageToken{MinRangeHash: shardLessThan + 1, MinTaskQueueId: minTaskQueueId})
-	}
-
-	if err != nil {
-		return nil, serviceerror.NewUnavailablef("error serializing nextPageToken:%v", err)
-	}
-
-	resp.NextPageToken = nextPageToken
-	return resp, nil
-}
-
-func (m *sqlTaskManagerV2) DeleteTaskQueue(
-	ctx context.Context,
-	request *persistence.DeleteTaskQueueRequest,
-) error {
-	nidBytes, err := primitives.ParseUUID(request.TaskQueue.NamespaceID)
-	if err != nil {
-		return serviceerror.NewUnavailable(err.Error())
-	}
-	tqId, tqHash := taskQueueIdAndHash(nidBytes, request.TaskQueue.TaskQueueName, request.TaskQueue.TaskQueueType, persistence.SubqueueZero)
-	result, err := m.Db.DeleteFromTaskQueuesV2(ctx, sqlplugin.TaskQueuesFilterV2{
-		RangeHash:   tqHash,
-		TaskQueueID: tqId,
-		RangeID:     &request.RangeID,
-	})
-	if err != nil {
-		return serviceerror.NewUnavailable(err.Error())
-	}
-	nRows, err := result.RowsAffected()
-	if err != nil {
-		return serviceerror.NewUnavailablef("rowsAffected returned error:%v", err)
-	}
-	if nRows != 1 {
-		return &persistence.ConditionFailedError{
-			Msg: fmt.Sprintf("delete failed: %v rows affected instead of 1", nRows),
-		}
-	}
-	return nil
-}
 func (m *sqlTaskManagerV2) CreateTasks(
 	ctx context.Context,
 	request *persistence.InternalCreateTasksRequest,
@@ -296,10 +63,12 @@ func (m *sqlTaskManagerV2) CreateTasks(
 	for i, v := range request.Tasks {
 		tqId, tqHash := idAndHash(v.Subqueue)
 		tasksRows[i] = sqlplugin.TasksRowV2{
-			RangeHash:    tqHash,
-			TaskQueueID:  tqId,
-			TaskID:       v.TaskId,
-			TaskPass:     v.TaskPass,
+			RangeHash:   tqHash,
+			TaskQueueID: tqId,
+			FairLevel: sqlplugin.FairLevel{
+				TaskPass: v.TaskPass,
+				TaskID:   v.TaskId,
+			},
 			Data:         v.Task.Data,
 			DataEncoding: v.Task.EncodingType.String(),
 		}
@@ -430,10 +199,10 @@ func lockTaskQueueV2(
 	tqId []byte,
 	oldRangeID int64,
 ) error {
-	rangeID, err := tx.LockTaskQueuesV2(ctx, sqlplugin.TaskQueuesFilterV2{
+	rangeID, err := tx.LockTaskQueues(ctx, sqlplugin.TaskQueuesFilter{
 		RangeHash:   tqHash,
 		TaskQueueID: tqId,
-	})
+	}, sqlplugin.MatchingTaskVersion2)
 	switch err {
 	case nil:
 		if rangeID != oldRangeID {
